+++ src/Stack/Docker.hs
@@ -61,7 +61,7 @@ import           System.Process.Run
 import           System.Process (CreateProcess(delegate_ctlc))
 import           Text.Printf (printf)
 
-#ifndef mingw32_HOST_OS
+#ifndef WINDOWS
 import           Control.Monad.Trans.Control (liftBaseWith)
 import           System.Posix.Signals
 #endif
@@ -247,6 +247,7 @@ runContainerAndExit modConfig
          ,concatMap (\(k,v) -> ["-e", k ++ "=" ++ v]) envVars
          ,concatMap sandboxSubdirArg sandboxSubdirs
          ,concatMap mountArg (dockerMount docker)
+         ,concatMap (\nv -> ["-e", nv]) (dockerEnv docker)
          ,case dockerContainerName docker of
             Just name -> ["--name=" ++ name]
             Nothing -> []
@@ -257,7 +258,7 @@ runContainerAndExit modConfig
          ,[cmnd]
          ,args])
      before
-#ifndef mingw32_HOST_OS
+#ifndef WINDOWS
      runInBase <- liftBaseWith $ \run -> return (void . run)
      oldHandlers <- forM (concat [[(sigINT,sigTERM) | not keepStdinOpen]
                                  ,[(sigTERM,sigTERM)]]) $ \(sigIn,sigOut) -> do
@@ -275,7 +276,7 @@ runContainerAndExit modConfig
                          ,["-a" | not (dockerDetach docker)]
                          ,["-i" | keepStdinOpen]
                          ,[containerID]]))
-#ifndef mingw32_HOST_OS
+#ifndef WINDOWS
      forM_ oldHandlers $ \(sig,oldHandler) ->
        liftIO $ installHandler sig oldHandler Nothing
 #endif
@@ -677,6 +678,7 @@ dockerOptsFromMonoid mproject stackRoot DockerOptsMonoid{..} = DockerOpts
   ,dockerContainerName = emptyToNothing dockerMonoidContainerName
   ,dockerRunArgs = dockerMonoidRunArgs
   ,dockerMount = dockerMonoidMount
+  ,dockerEnv = dockerMonoidEnv
   ,dockerDatabasePath =
      case dockerMonoidDatabasePath of
        Nothing -> stackRoot </> $(mkRelFile "docker.db")