+++ src/Stack/Exec.hs
@@ -15,7 +15,7 @@ import           Control.Monad.Trans.Control (MonadBaseControl)
 import           Stack.Types
 import           System.Process.Log
 
-#ifdef mingw32_HOST_OS
+#ifdef WINDOWS
 import           Control.Exception.Lifted
 import           Data.Streaming.Process (ProcessExitedUnsuccessfully(..))
 import           System.Exit
@@ -31,6 +31,7 @@ defaultEnvSettings = EnvSettings
     { esIncludeLocals = True
     , esIncludeGhcPackagePath = True
     , esStackExe = True
+    , esLocaleUtf8 = False
     }
 
 -- | Environment settings which do not embellish the environment
@@ -39,6 +40,7 @@ plainEnvSettings = EnvSettings
     { esIncludeLocals = False
     , esIncludeGhcPackagePath = False
     , esStackExe = False
+    , esLocaleUtf8 = False
     }
 
 -- | Execute a process within the Stack configured environment.
@@ -48,7 +50,7 @@ exec envSettings cmd0 args = do
     config <- asks getConfig
     menv <- liftIO (configEnvOverride config envSettings)
     $logProcessRun cmd0 args
-#ifdef mingw32_HOST_OS
+#ifdef WINDOWS
     e <- try (callProcess Nothing menv cmd0 args)
     liftIO $ case e of
         Left (ProcessExitedUnsuccessfully _ ec) -> exitWith ec