+++ src/Stack/Types/Config.hs
@@ -193,6 +193,8 @@ data EnvSettings = EnvSettings
     -- ^ include the GHC_PACKAGE_PATH variable
     , esStackExe :: !Bool
     -- ^ set the STACK_EXE variable to the current executable name
+    , esLocaleUtf8 :: !Bool
+    -- ^ set the locale to C.UTF-8
     }
     deriving (Show, Eq, Ord)
 
@@ -906,6 +908,7 @@ minimalEnvSettings =
     { esIncludeLocals = False
     , esIncludeGhcPackagePath = False
     , esStackExe = False
+    , esLocaleUtf8 = False
     }
 
 getWhichCompiler :: (MonadReader env m, HasEnvConfig env) => m WhichCompiler