+++ test/integration/tests/32-unlisted-module/files/src/Unlisted_FAIL.hs
@@ -1,0 +1,5 @@
+-- | Version of Unlisted with different export that causes failure to compile.
+module Unlisted where
+
+fooRenamed :: String
+fooRenamed = "foo"