+++ test/integration/tests/32-unlisted-module/files/src/Unlisted_OK.hs
@@ -1,0 +1,4 @@
+module Unlisted where
+
+foo :: String
+foo = "foo"